fix: remove default license type from  sbox create so config file can specify it
